Fix port numbers and rename files.

+++ src/tests/test_state_manager.h
@@ -48,7 +48,7 @@ public:
 
     bool register_data_service_apis(nuraft_mesg::Manager* messaging);
     static void fill_data_vec(nuraft_mesg::io_blob_list_t& cli_buf);
-    static uint32_t get_random_num();
+    static uint16_t get_random_num();
     static uint32_t get_server_counter();
     static void verify_data(sisl::io_blob const& buf);
